feat(permissions): Add support for separate REACT permission

+++ app/src/main/java/com/nextcloud/talk/chat/ChatActivity.kt
@@ -3977,6 +3977,10 @@ class ChatActivity :
     }
 
     override fun onClickReaction(chatMessage: ChatMessage, emoji: String) {
+        if (!participantPermissions.hasReactPermission()) {
+            Snackbar.make(binding.root, R.string.reaction_forbidden, Snackbar.LENGTH_LONG).show()
+            return
+        }
         VibrationUtils.vibrateShort(context)
         if (chatMessage.reactionsSelf?.contains(emoji) == true) {
             chatViewModel.deleteReaction(roomToken, chatMessage, emoji)
@@ -3995,7 +3999,7 @@ class ChatActivity :
             roomToken,
             chatMessage,
             conversationUser,
-            participantPermissions.hasChatPermission(),
+            participantPermissions.hasReactPermission(),
             ncApi
         ).show()
     }
@@ -4030,6 +4034,7 @@ class ChatActivity :
                 currentConversation,
                 isShowMessageDeletionButton(message),
                 participantPermissions.hasChatPermission(),
+                participantPermissions.hasReactPermission(),
                 spreedCapabilities
             ).show()
         }
